+++ src/app/(password-manager)/_layout.tsx
@@ -1,0 +1,5 @@
+import { Stack } from "expo-router";
+
+export default function PasswordManagerLayout() {
+  return <Stack screenOptions={{ headerShown: false }} />;
+}
