+++ src/app/_layout.tsx
@@ -4,12 +4,13 @@ import {
   isBiometricsAuthEnabled,
   unlockWithBiometricsApp,
 } from "@/libs/biometric";
+import FontAwesome6 from "@react-native-vector-icons/fontawesome6";
 import { NavigationBar } from "expo-navigation-bar";
-import { DarkTheme, DefaultTheme, Stack, ThemeProvider } from "expo-router";
+import { DarkTheme, DefaultTheme, Tabs, ThemeProvider } from "expo-router";
 import { StatusBar } from "expo-status-bar";
 import { useEffect, useRef, useState } from "react";
 import { AppState, Button, useColorScheme } from "react-native";
-import { PaperProvider } from "react-native-paper";
+import { PaperProvider, useTheme } from "react-native-paper";
 import { SafeAreaView } from "react-native-safe-area-context";
 
 export default function RootLayout() {
@@ -93,7 +94,42 @@ function AppContent() {
   return (
     <ThemeProvider value={isDarkScheme ? DarkTheme : DefaultTheme}>
       <PaperProvider>
-        <Stack screenOptions={{ headerShown: true }} />
+        <Tabs
+          screenOptions={{
+            headerShown: true,
+            tabBarActiveTintColor: useTheme().colors.tertiary,
+            headerTitleAlign: "center",
+          }}
+        >
+          <Tabs.Screen
+            name="(password-manager)"
+            options={{
+              title: "Passcodes",
+              tabBarIcon: ({ size, color }) => (
+                <FontAwesome6
+                  name="key"
+                  size={size}
+                  color={color}
+                  iconStyle="solid"
+                />
+              ),
+            }}
+          />
+          <Tabs.Screen
+            name="(settings)"
+            options={{
+              title: "Settings",
+              tabBarIcon: ({ size, color }) => (
+                <FontAwesome6
+                  name="gear"
+                  size={size}
+                  color={color}
+                  iconStyle="solid"
+                />
+              ),
+            }}
+          />
+        </Tabs>
       </PaperProvider>
     </ThemeProvider>
   );
